ModerateExistingFiles: Process single file without job queue
The job queue implementation makes debugging issues with a single file
more complicated than it needs to be. This patch bypasses the job queue
implementation when the `--file-name` argument is supplied to the
maintenance script.

To test:

 php maintenance/run.php ./extensions/MediaModeration/maintenance/ModerateExistingFiles.php --file-name=Test.png

If --file-name is not specified, then the maintenance script will
enqueue jobs for batch processing.

Misc:

* Use dependency injection for the JobQueueGroup class in ModerateExistingFilesHelper

Bug: T336370
Change-Id: I098756dbaff79b6ff85ccf47e14e9eb8eadd4465