Remove broken exception handling
First of all, it's unclear if the exception in question is checked and
should be caught. But regardless of that, if the exception is thrown,
the current code will not behave correctly:
- If the exception is thrown in the first loop iteration, it will pass a
  null $file to processFile, which will throw a TypeError because the
  parameter is typehinted as LocalFile.
- If the exception is thrown in a later iteration, it will reuse the
  same $file variable as the last iteration.

Maybe the intention was to have a `continue` after printing the
exception details, but since this code seems to have been working fine
so far, the `try...catch` appears unnecessary.

Bug: T328220
Change-Id: I40b42cd8c4e7a2990e00f8258e2aa0fd3e96b2ff