phpunit: Fix UpdateMetricsTest to not rely on exact flush threshold
At the moment, Maintenance::output() calls StatsFactory::flush once
the buffer reaches a certain threshold. It is also called automatically
after any database transaction commits.

Test test relied on accessing the internal StatsCache and finding the
metric object *before* it is flushed.

Switch to StatsFactory::newUnitTestingHelper which provides a higher
confidence assertion (full metric, value, labels, order, and no
other unexpected metrics).

While at it, de-sugar the data provider for now to simplify the code,
since the provider did not control the test in any way, the parameter
were only related to the output, not the scenario itself, so if there
were more than one, they would have to have the same output.

If other cases are needed in the future, I suggest a data provider
that returns two strings, an output string, and a stats string.

Bug: T381042
Change-Id: Ib490666ffa4bf87234abf156bd5ae1a5ad9282db